Proposal for file structure organization
