--- android/src/main/java/com/swmansion/rnscreens/modals/formsheet/FormSheetContentSizeChangeDelegate.kt
@@ -1,5 +1,0 @@
-package com.swmansion.rnscreens.modals.formsheet
-
-interface FormSheetContentSizeChangeDelegate {
-    fun onContentHeightChanged(newHeight: Int)
-}
