--- android/src/main/java/com/swmansion/rnscreens/modals/formsheet/FormSheetContentSizeChangeProvider.kt
@@ -1,5 +1,0 @@
-package com.swmansion.rnscreens.modals.formsheet
-
-interface FormSheetContentSizeChangeProvider {
-    fun setContentSizeChangeDelegate(delegate: FormSheetContentSizeChangeDelegate?)
-}
