--- android/src/main/java/com/swmansion/rnscreens/modals/formsheet/FormSheetDialogEventEmitter.kt
@@ -1,11 +1,0 @@
-package com.swmansion.rnscreens.modals.formsheet
-
-import com.swmansion.rnscreens.common.event.ViewAppearanceEventEmitter
-
-internal interface FormSheetDialogEventEmitter : ViewAppearanceEventEmitter {
-    fun emitOnNativeDismissEvent()
-
-    fun emitOnNativeDismissPreventedEvent()
-
-    fun emitOnDetentChanged(index: Int)
-}
