--- android/src/main/java/com/swmansion/rnscreens/modals/formsheet/FormSheetHostEventEmitter.kt
@@ -1,58 +1,0 @@
-package com.swmansion.rnscreens.modals.formsheet
-
-import com.facebook.react.bridge.ReactContext
-import com.swmansion.rnscreens.common.event.BaseEventEmitter
-
-internal class FormSheetHostEventEmitter(
-    reactContext: ReactContext,
-    viewTag: Int,
-) : BaseEventEmitter(reactContext, viewTag),
-    FormSheetDialogEventEmitter {
-    override fun emitOnNativeDismissEvent() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetNativeDismissEvent(surfaceId, viewTag),
-        )
-    }
-
-    fun emitOnSyncFlushEvent() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetSyncFlushEvent(surfaceId, viewTag),
-        )
-    }
-
-    override fun emitOnWillAppear() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetWillAppearEvent(surfaceId, viewTag),
-        )
-    }
-
-    override fun emitOnDidAppear() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetDidAppearEvent(surfaceId, viewTag),
-        )
-    }
-
-    override fun emitOnWillDisappear() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetWillDisappearEvent(surfaceId, viewTag),
-        )
-    }
-
-    override fun emitOnDidDisappear() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetDidDisappearEvent(surfaceId, viewTag),
-        )
-    }
-
-    override fun emitOnDetentChanged(index: Int) {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetDetentChangedEvent(surfaceId, viewTag, index),
-        )
-    }
-
-    override fun emitOnNativeDismissPreventedEvent() {
-        reactEventDispatcher.dispatchEvent(
-            FormSheetNativeDismissPreventedEvent(surfaceId, viewTag),
-        )
-    }
-}
