--- android/src/main/java/com/swmansion/rnscreens/modals/formsheet/FormSheetWillAppearEvent.kt
@@ -1,29 +1,0 @@
-package com.swmansion.rnscreens.modals.formsheet
-
-import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.WritableMap
-import com.facebook.react.uimanager.events.Event
-import com.swmansion.rnscreens.common.event.NamingAwareEventType
-
-class FormSheetWillAppearEvent(
-    surfaceId: Int,
-    viewId: Int,
-) : Event<FormSheetWillAppearEvent>(surfaceId, viewId),
-    NamingAwareEventType {
-    override fun getEventName(): String = EVENT_NAME
-
-    override fun getEventRegistrationName() = EVENT_REGISTRATION_NAME
-
-    override fun getEventData(): WritableMap? = Arguments.createMap()
-
-    override fun canCoalesce(): Boolean = false
-
-    companion object : NamingAwareEventType {
-        const val EVENT_NAME = "topWillAppear"
-        const val EVENT_REGISTRATION_NAME = "onWillAppear"
-
-        override fun getEventName() = EVENT_NAME
-
-        override fun getEventRegistrationName() = EVENT_REGISTRATION_NAME
-    }
-}
