+++ android/src/main/java/com/swmansion/rnscreens/modals/formsheet/shared/contracts/FormSheetContentSizeChangeDelegate.kt
@@ -1,0 +1,5 @@
+package com.swmansion.rnscreens.modals.formsheet.shared.contracts
+
+interface FormSheetContentSizeChangeDelegate {
+    fun onContentHeightChanged(newHeight: Int)
+}
