+++ android/src/main/java/com/swmansion/rnscreens/modals/formsheet/shared/contracts/FormSheetContentSizeChangeProvider.kt
@@ -1,0 +1,5 @@
+package com.swmansion.rnscreens.modals.formsheet.shared.contracts
+
+interface FormSheetContentSizeChangeProvider {
+    fun setContentSizeChangeDelegate(delegate: FormSheetContentSizeChangeDelegate?)
+}
