+++ android/src/main/java/com/swmansion/rnscreens/modals/formsheet/shared/contracts/FormSheetDialogEventEmitter.kt
@@ -1,0 +1,11 @@
+package com.swmansion.rnscreens.modals.formsheet.shared.contracts
+
+import com.swmansion.rnscreens.common.event.ViewAppearanceEventEmitter
+
+internal interface FormSheetDialogEventEmitter : ViewAppearanceEventEmitter {
+    fun emitOnNativeDismissEvent()
+
+    fun emitOnNativeDismissPreventedEvent()
+
+    fun emitOnDetentChanged(index: Int)
+}
